Upload Assignemenst one, 2 and 3
I had problems with the syncronizing my commits as the files kept getting corrupted, Innitially I had commited part 1 and half of part 2 but the the files crashed and I had to delete my repo and start all over.

Assignment 1 went smoothly colours and font where changed, as I added the second page I did have to play with the fonts more as sometimes they were not visible. I had trouble with the button as  when doing the same style as the document I would get the "button" but with no padding and no "box" for this reason I opted for another method that did work and that  fulfilled othe sizing. The images gave me the most trouble as files kept breaking and it was then not possible to read their origin. Funny note I was trying to use html code in CSS that cost me some time to realize why some elements where not loading. Less funny I maded this assignment more times then should be needed I am sorry for not having more commits, but it was my own sweat and almost tears, over some basic code :P

- Assignment one addapted fonts and background, new background image
- Extra page for  work experience
- Image gallery that extends.
! missing to re-add the linking between pages